# Speed up Formal Verification