--- a/.gitignore
+++ b/.gitignore
@@ -1,1 +1,3 @@
 /uhdm-tests/opentitan/opentitan-parsing-test/build/
+__pycache__
+/build