| Test                   | Time before | Time now |
|:-----------------------|------------:|---------:|
| UHMD-integration-tests |       47:16 |    13:05 |
| sv2v                   |     1:07:48 |    16:16 |
| yosys                  |     1:14:59 |    23:30 |

This also adds generic Python module `run_command.py` which allows running processes.
Extra features when compared to `subprocess.run()`:
- Allow specifying time limit, address space limit, `oom_score_adj`
- Returns max RSS used by child process, and system/user/wall time.

Log example: https://github.com/antmicro/yosys-systemverilog/actions/runs/4597493625/jobs/8120398417#step:9:295